Misc improvements and added depth control tests
