+++ catkin_ws/src/aquadrone_sim_demos/launch/empty.launch
@@ -11,7 +11,7 @@
     </include>
     
     <include file="$(find aquadrone_description)/launch/spawn_sub.launch">
-        <arg name="model" value="aquadrone_v2"/>
+        <arg name="model" value="aquadrone_$(arg model_name)"/>
     </include>
 
     <include file="$(find thruster_control)/launch/$(arg control_file)"/>
