+++ catkin_ws/src/thruster_control/nodes/sim_thruster_controller.py
@@ -10,7 +10,7 @@ from thruster_control.thruster_interfaces import SimulatedThrusterInterface
 if __name__ == "__main__":
     rospy.init_node('sim_thruster_controller')
 
-    thrusters = [SimulatedThrusterInterface(i) for i in range(0, 6)]
+    thrusters = [SimulatedThrusterInterface('aquadrone_v1', i) for i in range(0, 6)]
 
     tm = ThrusterCollectionManager(thrusters)
     while not rospy.is_shutdown():
