+++ catkin_ws/src/thruster_control/test/test_control.test
@@ -1,0 +1,3 @@
+<launch>
+  <test test-name="test_controls" pkg="thruster_control" type="test_depth_control.py" />
+</launch>
